moved type support in own file
